% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addBiotypeClasses.R
\name{addBiotypeClasses}
\alias{addBiotypeClasses}
\title{use data(repeat_biotypes) and/or data(ensembl_biotypes) to classify biotypes}
\usage{
addBiotypeClasses(txdblite)
}
\arguments{
\item{txdblite}{A txdblite object (should have metadata on its sqlite file)}
}
\description{
use data(repeat_biotypes) and/or data(ensembl_biotypes) to classify biotypes
}

