% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrayControlDbLite.R
\name{ArrayControlDbLite}
\alias{ArrayControlDbLite}
\title{Constructor for an ArrayControlDbLite object.}
\usage{
ArrayControlDbLite(x, ...)
}
\arguments{
\item{x}{the name of the sqlite file}

\item{...}{additional arguments related to ArrayControl annotation database.}
}
\value{
an ErccDbLite object.
}
\description{
Constructor for an ArrayControlDbLite object.
}
