% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chrs.R
\name{chrs}
\alias{chrs}
\title{helper function for findDupes}
\usage{
chrs(fastaFile)
}
\arguments{
\item{fastaFile}{the target FASTA}
}
\value{
a data.frame with the contig entry names and the FASTA file of origin
}
\description{
helper function for findDupes
}

