% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnnotationPackage.R
\name{createAnnotationPackage}
\alias{createAnnotationPackage}
\title{create an annotation package for a FASTA file (try to figure out what kind)}
\usage{
createAnnotationPackage(fastaFile, ...)
}
\arguments{
\item{fastaFile}{the filename}
}
\value{
the name of the annotation package, or NULL if uncertain how to do it
}
\description{
create an annotation package for a FASTA file (try to figure out what kind)
}

