% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsDbLite.R
\name{EnsDbLite}
\alias{EnsDbLite}
\title{Constructor for an EnsDbLite object.}
\usage{
EnsDbLite(x, ...)
}
\arguments{
\item{x}{the name of the sqlite file}

\item{...}{any additional arguments relating to ENSEMBL annotation database.}
}
\value{
an EnsDbLite object.
}
\description{
Constructor for an EnsDbLite object.
}

