% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensDbLiteExonsFromGtf.R
\name{ensDbLiteExonsFromGtf}
\alias{ensDbLiteExonsFromGtf}
\title{as the name suggests, this fetches transcript-to-exon mappings from a GTF.}
\usage{
ensDbLiteExonsFromGtf(ensdb, gtf)
}
\arguments{
\item{ensdb}{an (existing) EnsDbLite object (with con, etc.)}

\item{gtf}{the file name for the GTF file (perhaps compressed)}
}
\description{
as the name suggests, this fetches transcript-to-exon mappings from a GTF.
}

