% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensembl_biotypes}
\alias{ensembl_biotypes}
\title{Structure for annotating transcript biotypes from ENSEMBL}
\format{An object of class \code{data.frame} with 48 rows and 2 columns.}
\source{
\url{http://www.ensembl.org/}
}
\usage{
ensembl_biotypes
}
\description{
holds 48 transcript biotypes and 6 levels of classes for Homo Sapiens.
}
\keyword{datasets}

