% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensemblToReactome}
\alias{ensemblToReactome}
\title{Data frame of ENSEMBL transcript IDs and corresponding reactome pathway IDs, Reactome URL pathway-browser, reactome pathway name, and organism for numerous species including Mus musculus, Homo Sapiens and many more.  This is used for enrichment activation analysis.}
\format{A data frame full of reactome data
  \itemize{
   \item code: names of ENSEMBL IDs of transcripts and genes
   \item term: Reactome ID pathway
  \item URL: Reactome Pathway URL
 \item name: Pathway name
\item from: Pathway Origin
\item organism: organism of pathway
}}
\source{
\url{www.reactome.org}
}
\usage{
ensemblToReactome
}
\description{
Data frame of ENSEMBL transcript IDs and corresponding reactome pathway IDs, Reactome URL pathway-browser, reactome pathway name, and organism for numerous species including Mus musculus, Homo Sapiens and many more.  This is used for enrichment activation analysis.
}
\keyword{datasets}

