% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ERCC_annotated}
\alias{ERCC_annotated}
\title{Data frame of External RNA-Control-Consortium (ERCC) Ambion Spike-Ins Concentration Mixes}
\format{An object of class \code{data.frame} with 92 rows and 3 columns.}
\source{
\url{https://www.thermofisher.com/order/catalog/product/4456740}
}
\usage{
ERCC_annotated
}
\description{
Data frame of External RNA-Control-Consortium (ERCC) Ambion Spike-Ins Concentration Mixes
}
\keyword{datasets}
