% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErccDbLite.R
\name{ErccDbLite}
\alias{ErccDbLite}
\title{Constructor for an ErccDbLite object.}
\usage{
ErccDbLite(x, ...)
}
\arguments{
\item{x}{the name of the sqlite file}

\item{...}{additional arguments related to ERCC annotation database.}
}
\value{
an ErccDbLite object.
}
\description{
Constructor for an ErccDbLite object.
}

