% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchFromEnsembl.R
\name{fetchFromEnsembl}
\alias{fetchFromEnsembl}
\title{a wrapper for biomaRt::getBM}
\usage{
fetchFromEnsembl(cols, idtype, IDs = NULL, dataset)
}
\arguments{
\item{cols}{attributes to fetch}

\item{idtype}{ID type filter (or a list of filters, more rarely)}

\item{IDs}{the IDs to try and fetch (NULL if filters is a list)}

\item{dataset}{the dataset to use}
}
\value{
a data.frame, if all goes well
}
\description{
a wrapper for biomaRt::getBM
}

