% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationType.R
\name{getAnnotationType}
\alias{getAnnotationType}
\title{(try to) figure out what type of annotation package to create}
\usage{
getAnnotationType(fastaFile)
}
\arguments{
\item{fastaFile}{a fasta file of interest}
}
\description{
(try to) figure out what type of annotation package to create
}
