% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDupeSeq.R
\name{getDupeSeq}
\alias{getDupeSeq}
\title{helper function for findDupes}
\usage{
getDupeSeq(duperow)
}
\arguments{
\item{duperow}{the row}
}
\value{
the sequence for that duplicated entry
}
\description{
helper function for findDupes
}
