% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGCcontent.R
\name{getGCcontent}
\alias{getGCcontent}
\alias{GCcontent}
\title{Get the GC content of every contig in a FASTA.}
\usage{
getGCcontent(fastaFile)

GCcontent(seqs)
}
\arguments{
\item{fastaFile}{the FASTA file.}

\item{seqs}{a DNAStringSet, RNAStringSet, or similar}
}
\value{
a GRanges with a gc_content column.

a vector of GC proportions
}
\description{
Get the GC content of every contig in a FASTA.
}
\section{Functions}{
\itemize{
\item \code{GCcontent}: helper function
}}

