% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHugoCache.R
\name{getHugoCache}
\alias{getHugoCache}
\title{HUGO transcript and gene names for ENSEMBL IDs of the corresponding type}
\usage{
getHugoCache(useCache = TRUE)
}
\arguments{
\item{useCache}{boolean (TRUE, and should stay that way, save for updates)}
}
\value{
a list of HUGO names associated with ENSEMBL transcript and gene IDs
}
\description{
HUGO transcript and gene names for ENSEMBL IDs of the corresponding type
}
