% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMostRecentCacheBuild.R
\name{getMostRecentCacheBuild}
\alias{getMostRecentCacheBuild}
\title{Indicate what the most recent ENSEMBL build is that we have cached (manually)}
\usage{
getMostRecentCacheBuild()
}
\value{
build           a numeric indicator for the most recent cached build
}
\description{
Indicate what the most recent ENSEMBL build is that we have cached (manually)
}
