% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReactomeUrl.R
\name{getReactomeUrl}
\alias{getReactomeUrl}
\title{Function that lets us avoid storing a redundant pile of URLs in a cache}
\usage{
getReactomeUrl(term)
}
\arguments{
\item{term}{the Reactome term to look up}
}
\value{
url  the URL to look it up on the Reactome website
}
\description{
Function that lets us avoid storing a redundant pile of URLs in a cache
}
