% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hugoCache}
\alias{hugoCache}
\title{A list of HUGO Gene Symbols and associated ENSEMBL Transcripts and Gene ID}
\format{A list of two, 
\itemize{
 \item transcript: a list of ENSEMBL Transcript IDs and associated HUGO Gene Symbols
 \item gene: a list of ENSEMBL Gene IDs and associated HUGO Gene Symbols.
}}
\source{
\url{www.genenames.org}
}
\usage{
hugoCache
}
\description{
A list of HUGO Gene Symbols and associated ENSEMBL Transcripts and Gene ID
}
\keyword{datasets}
