% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeEnsDbLitePkg.R
\name{makeEnsDbLitePkg}
\alias{makeEnsDbLitePkg}
\title{create an EnsDbLite package from sqlite file (usually via ensDbLiteFromFasta)}
\usage{
makeEnsDbLitePkg(ensdblitefile, author, email, version = "1.0",
  destDir = ".", ...)
}
\arguments{
\item{ensdblitefile}{the sqlite filename}

\item{author}{whose fault this is}

\item{email}{adding maintainer's email}

\item{version}{version of the package (default is "1.0")}

\item{destDir}{where to put the new package directory (".")}
}
\value{
the name of the package
}
\description{
create an EnsDbLite package from sqlite file (usually via ensDbLiteFromFasta)
}

