% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFI.R
\name{mapFI}
\alias{mapFI}
\title{map Reactome functional interactions (FIs)}
\usage{
mapFI(species)
}
\arguments{
\item{symbols}{the identifiers to map}

\item{type}{what type of identifier are these? (gene)}
}
\value{
a data.frame mapping IDs to HUGO symbols or transcript names
}
\description{
map Reactome functional interactions (FIs)
}

