% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapHugo.R
\name{mapHugo}
\alias{mapHugo}
\title{map genes or transcripts to updated HUGO symbols}
\usage{
mapHugo(IDs, type = c("gene", "transcript"))
}
\arguments{
\item{IDs}{the identifiers to map}

\item{type}{what type of identifier are these? (gene)}
}
\value{
a data.frame mapping IDs to HUGO symbols or transcript names
}
\description{
map genes or transcripts to updated HUGO symbols
}

