% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapReactomeSets.R
\name{mapReactomeSets}
\alias{mapReactomeSets}
\title{this creates a reactomePathway set in terms of gene names from a reactomeSet which is a list in terms of reactomeId per ENSEMBL entry.  *This Function is deprecated*}
\usage{
mapReactomeSets(reactomePathways, reactomeSets)
}
\arguments{
\item{reactomePathways}{a cache rda object list with all the reactome Ids listing description of pathways}

\item{reactomeSets}{a list of reactome Ids per ENSEMBL gene / transcript set}
}
\value{
a list of reactomePathways per associated gene name
}
\description{
this creates a reactomePathway set in terms of gene names from a reactomeSet which is a list in terms of reactomeId per ENSEMBL entry.  *This Function is deprecated*
}
