% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGenesToReactome.R
\name{mapToReactome}
\alias{mapToReactome}
\title{map genes or transcripts to Reactome pathways}
\usage{
mapToReactome(IDs, type = c("gene", "transcript"))
}
\arguments{
\item{IDs}{the identifiers to map}

\item{type}{what type of identifier are these? (gene)}
}
\value{
a data.frame mapping IDs to Reactome IDs and pathways
}
\description{
map genes or transcripts to Reactome pathways
}

