% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MitxDbLite.R
\name{MitxDbLite}
\alias{MitxDbLite}
\title{Constructor for an MitxDbLite (miTranscriptome) object.}
\usage{
MitxDbLite(x, ...)
}
\arguments{
\item{x}{the name of the sqlite file}
}
\value{
an MitxDbLite object.
}
\description{
Constructor for an MitxDbLite (miTranscriptome) object.
}

