% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitxDbLiteFromFasta.R
\name{mitxDbLiteFromFasta}
\alias{getEntrezIDs}
\alias{getSymbols}
\alias{mitxDbLiteFromFasta}
\alias{mitxDbLiteMetadata}
\title{Function to create an MitxDbLite object (a stripped-down SQLite db) from
a specified miTranscriptome FASTA file for a given version and type(s).
Note that this is shockingly easier than the same process for the GTF...}
\usage{
mitxDbLiteFromFasta(fastaFile, verbose = TRUE)

getEntrezIDs(gxs, organism)

getSymbols(gxs, organism)

mitxDbLiteMetadata(packageName, genomeVersion, sourceFile)
}
\arguments{
\item{fastaFile}{the FASTA file to collate into a MitxDbLite instance}

\item{verbose}{make a lot of noise? (TRUE)}

\item{gxs}{a GRanges of genes}

\item{organism}{what kind of organism these genes are from}

\item{packageName}{the name of the annotation package to be built}

\item{genomeVersion}{name of genome assembly for coordinates, e.g. "GRCh38"}

\item{sourceFile}{name of FASTA file(s) whence it was built, as a string}

\item{gxs}{a GRanges of genes}
}
\value{
entrez_id values for the genes, where found

symbols for the genes, where found

a data.frame of metadata suitable for cramming into the database
}
\description{
Function to create an MitxDbLite object (a stripped-down SQLite db) from
a specified miTranscriptome FASTA file for a given version and type(s).
Note that this is shockingly easier than the same process for the GTF...
}
\section{Functions}{
\itemize{
\item \code{getEntrezIDs}: add EntrezGene IDs for Ensembl genes

\item \code{getSymbols}: add symbols for Ensembl genes

\item \code{mitxDbLiteMetadata}: create metadata for an MitxDbLite instance
}}

