% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_repeat_biotypes}
\alias{mouse_repeat_biotypes}
\title{A Data frame of repetitive elements relating to Mus Musculus}
\format{A Data frame filled with mouse repeats
\itemize{
  \item name: names of repeat elements
  \item tx_biotype: transcript biotype of given repeat
  \item gene_biotype: gene biotype
  \item class: class of transcript
  }}
\source{
\url{www.girinst.org}
}
\usage{
mouse_repeat_biotypes
}
\description{
A Data frame of repetitive elements relating to Mus Musculus
}
\keyword{datasets}

