% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactomeCache}
\alias{reactomeCache}
\title{A list of ENSEMBL Transcripts and associated Reactome Pathway IDs}
\format{An object of class \code{list} of length 6.}
\source{
\url{www.reactome.org}
}
\usage{
reactomeCache
}
\description{
A list of ENSEMBL Transcripts and associated Reactome Pathway IDs
}
\keyword{datasets}

