% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactomePathways}
\alias{reactomePathways}
\title{A list of Reactome IDs and associated names for each reactome ID}
\format{An object of class \code{list} of length 21170.}
\source{
\url{www.reactome.org}
}
\usage{
reactomePathways
}
\description{
A list of Reactome IDs and associated names for each reactome ID
}
\keyword{datasets}
