% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RepDbLite.R
\name{RepDbLite}
\alias{RepDbLite}
\title{Constructor for an RepDbLite object.}
\usage{
RepDbLite(x, ...)
}
\arguments{
\item{x}{the name of the sqlite file}
}
\value{
a RepDbLite object.
}
\description{
Constructor for an RepDbLite object.
}

