% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repDbLiteFromMouseFasta.R
\name{repDbLiteFromMouseFasta}
\alias{repDbLiteFromMouseFasta}
\title{functions to create a RepDbLite object (a stripped-down SQLite db) from
a specified RepBase FASTA file for a given version, species, and type(s). this is the mouse rep biotype to be used.}
\usage{
repDbLiteFromMouseFasta(fastaFile, verbose = TRUE)
}
\arguments{
\item{fastaFile}{the FASTA file to collate into a RepDbLite instance}

\item{verbose}{make a lot of noise? (TRUE)}
}
\description{
functions to create a RepDbLite object (a stripped-down SQLite db) from
a specified RepBase FASTA file for a given version, species, and type(s). this is the mouse rep biotype to be used.
}

