% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repDbLiteFromFasta.R
\name{repDbLiteMetadata}
\alias{repDbLiteMetadata}
\title{create metadata for a RepDbLite instance}
\usage{
repDbLiteMetadata(packageName, sourceFile)
}
\arguments{
\item{packageName}{the name of the annotation package to be built}

\item{sourceFile}{name of FASTA file(s) whence it was built, as a string}
}
\value{
a data.frame of metadata suitable for cramming into the database
}
\description{
create metadata for a RepDbLite instance
}
