% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{repeat_biotypes}
\alias{repeat_biotypes}
\title{A Data frame of repetitive elements relating to Homo Sapiens}
\format{A Data frame filled with human repeats
\itemize{
  \item name: names of repeat elements
  \item tx_biotype: transcript biotype of given repeat
  \item gene_biotype: gene biotype
  \item class: class of transcript
  }}
\source{
\url{www.girinst.org}
}
\usage{
repeat_biotypes
}
\description{
A Data frame of repetitive elements relating to Homo Sapiens
}
\keyword{datasets}
