% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikeinDbLiteFromFasta.R
\name{spikeinDbLiteFromFasta}
\alias{spikeinDbLiteFromFasta}
\title{create an general spike in annotation DB from the FASTA file
NOTE: we probably shouldn't even bother exporting this.}
\usage{
spikeinDbLiteFromFasta(fastaFile, verbose = TRUE, dryRun = FALSE)
}
\arguments{
\item{fastaFile}{a fasta file containing ERCC sequence information}

\item{verbose}{boolean if true will print process messages}

\item{dryRun}{boolean if false a sql-lite data base is saved, if true, no database is saved.}
}
\description{
create an general spike in annotation DB from the FASTA file
NOTE: we probably shouldn't even bother exporting this.
}
