% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{supportedOrganismsForTxDbLite}
\alias{supportedOrganismsForTxDbLite}
\title{A Data Frame of all supported organisms and supported annotations available by TxDbLite}
\format{A Data Frame of information of supported species
\itemize{
\item name: names of species
\item abbreviation: supported species abbreviation
\item package: associated package for given species
\item keytype: species key
\item symbol: species symbol
\item txpre: transcrpit prefix for species
\item gxpre: gene prefix
\item reactome: reactome abbreviation
}}
\source{
\url{www.ensembl.org}
}
\usage{
supportedOrganismsForTxDbLite
}
\description{
A Data Frame of all supported organisms and supported annotations available by TxDbLite
}
\keyword{datasets}

