% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{TxDbLite-class}
\alias{TxDbLite-class}
\title{A parent class for lightweight Ensembl and RepBase/RepeatMasker annotations}
\description{
A parent class for lightweight Ensembl and RepBase/RepeatMasker annotations
}

