% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\title{utils}
\name{utils}
\alias{createAnnotationPackage}
\alias{getAnnotationType}
\alias{getFastaStub}
\usage{
getFastaStub(fastaFile)
getAnnotationType(fastaFile)
createAnnotationPackage(fastaFile, author)
}
\arguments{
\item{fastaFile}{the filename}

\item{author}{the author (with email address)}

\item{fastaFile}{the filename}
}
\value{
a string

the name of the annotation package, or NULL if uncertain how to do it
}
\section{Functions}{
\itemize{
\item \code{getFastaStub}: get the "stub" of a FASTA filename (no .fa, no .fasta, no .gz)

\item \code{getAnnotationType}: figure out what type of annotation package to create

\item \code{createAnnotationPackage}: create an annotation package for a FASTA file (try to figure out what kind)
}}

