% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EnsDbLite.R
\name{EnsDbLite}
\alias{EnsDbLite}
\title{Constructor for an EnsDbLite object.}
\usage{
EnsDbLite(x)
}
\arguments{
\item{x}{the name of the sqlite file}
}
\value{
an EnsDbLite object.
}
\description{
Constructor for an EnsDbLite object.
}

