% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ErccDbLite.R
\name{ErccDbLite}
\alias{ErccDbLite}
\title{Constructor for an ErccDbLite object.}
\usage{
ErccDbLite(x, ...)
}
\arguments{
\item{x}{the name of the sqlite file}
}
\value{
an ErccDbLite object.
}
\description{
Constructor for an ErccDbLite object.
}

