% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/erccDbLiteFromFasta.R
\name{erccDbLiteFromFasta}
\alias{erccDbLiteFromFasta}
\title{create an ERCC annotation DB from the FASTA file}
\usage{
erccDbLiteFromFasta(fastaFile, verbose = TRUE)
}
\description{
create an ERCC annotation DB from the FASTA file
}

