% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReactomeCache.R
\name{getReactomeCache}
\alias{getReactomeCache}
\title{Pulling Reactome mappings through biomaRt takes time. So we try not to do it.
More specifically, try to only pull updates, and then only for new builds.}
\usage{
getReactomeCache(species = "Homo sapiens", build = 84)
}
\arguments{
\item{species}{species for which to get mappings (Homo sapiens)}

\item{build}{numeric ENSEMBL build (84 by default)}
}
\value{
a list of REACTOME terms mapped to ENSEMBL IDs (both tx and gene)
}
\description{
Pulling Reactome mappings through biomaRt takes time. So we try not to do it.
More specifically, try to only pull updates, and then only for new builds.
}

