% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indexIfNoneFound.R
\name{indexIfNoneFound}
\alias{indexIfNoneFound}
\title{helper function for findDupes}
\usage{
indexIfNoneFound(fastaFile)
}
\arguments{
\item{fastaFile}{the name of the FASTA file

return   the index, invisibly}
}
\description{
helper function for findDupes
}

