% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapHugo.R
\name{mapHugo}
\alias{mapHugo}
\title{map genes or transcripts to updated HUGO symbols}
\usage{
mapHugo(IDs, useCache = TRUE, byType = c("transcript", "gene"))
}
\arguments{
\item{IDs}{the identifiers to map}

\item{useCache}{boolean: use the cache? (TRUE, and should stay that way)}

\item{byType}{gene, or transcript}
}
\value{
a list mapping IDs to HUGO symbols or HUGO transcript names
}
\description{
map genes or transcripts to updated HUGO symbols
}
\examples{
mapHugo("ENST00000003084")

}

