% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapToReactome.R
\name{mapToReactome}
\alias{mapToReactome}
\title{map transcripts or genes to Reactome pathways, using a cache if possible}
\usage{
mapToReactome(IDs, type = c("transcript", "gene"), species = "Homo sapiens",
  build = 84, asNames = FALSE, asUrls = FALSE, useCache = TRUE)
}
\arguments{
\item{IDs}{the identifiers to map}

\item{type}{transcript- or gene-level mappings? (default: transcript)}

\item{species}{what species are the IDs from? (default:Homo sapiens)}

\item{build}{what build the IDs are from? (default:84)}

\item{asNames}{boolean, convert the terms to full pathway names? (FALSE)}

\item{asUrls}{boolean, convert terms to Reactome website URLs? (FALSE)}

\item{useCache}{use the cache? (TRUE; do not change unless maintaining)}
}
\value{
a list of mappings for the supplied IDs
}
\description{
map transcripts or genes to Reactome pathways, using a cache if possible
}
\examples{
 mapToReactome("ENSG00000234258", type="gene")
           mapToReactome("ENST00000387459", asNames=TRUE)

}

