% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepDbLite.R
\name{RepDbLite}
\alias{RepDbLite}
\title{Constructor for an RepDbLite object.}
\usage{
RepDbLite(x, ...)
}
\arguments{
\item{x}{the name of the sqlite file}

\item{...}{additional arguments related to Repetitive annotation database.}
}
\value{
a RepDbLite object.
}
\description{
Constructor for an RepDbLite object.
}
