% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repDbLiteFromFasta.R
\name{repDbLiteFromFasta}
\alias{repDbLiteFromFasta}
\alias{repDbLiteMetadata}
\title{functions to create a RepDbLite object (a stripped-down SQLite db) from
a specified RepBase FASTA file for a given version, species, and type(s)}
\usage{
repDbLiteFromFasta(fastaFile, verbose = TRUE)

repDbLiteMetadata(packageName, sourceFile)
}
\arguments{
\item{fastaFile}{the FASTA file to collate into a RepDbLite instance}

\item{verbose}{make a lot of noise? (TRUE)}

\item{packageName}{the name of the annotation package to be built}

\item{sourceFile}{name of FASTA file(s) whence it was built, as a string}
}
\value{
a data.frame of metadata suitable for cramming into the database
}
\description{
functions to create a RepDbLite object (a stripped-down SQLite db) from
a specified RepBase FASTA file for a given version, species, and type(s)
}
\section{Functions}{
\itemize{
\item \code{repDbLiteMetadata}: create metadata for a RepDbLite instance
}}

