% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repDbLiteFromFasta.R
\name{repDbLiteFromFasta}
\alias{repDbLiteFromFasta}
\title{functions to create a RepDbLite object (a stripped-down SQLite db) from
specified RepBase FASTA files for a given version, species, and type(s)}
\usage{
repDbLiteFromFasta(organism, fastaFiles, version)
}
\arguments{
\item{organism}{the organism involved (Homo_sapiens, Mus_musculus, etc.)}

\item{fastaFiles}{the FASTA files to collate into a RepDbLite instance}

\item{version}{the RepBase build (e.g. 20_07)}
}
\description{
functions to create a RepDbLite object (a stripped-down SQLite db) from
specified RepBase FASTA files for a given version, species, and type(s)
}

