% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R, R/Methods.R
\docType{class}
\name{TxDbLite-class}
\alias{TxDbLite-class}
\alias{genesBy,TxDbLite-method}
\alias{transcriptsBy,TxDbLite-method}
\title{A parent class for lightweight Ensembl and RepBase/RepeatMasker annotations}
\usage{
\S4method{transcriptsBy}{TxDbLite}(x, by = c("gene", "promoter", "tx_biotype",
  "gene_biotype", "biotype_class"))

\S4method{genesBy}{TxDbLite}(x, by = c("gene_biotype", "biotype_class"))
}
\arguments{
\item{x}{the TxDbLite instance}

\item{by}{how to split the transcripts}

\item{x}{the TxDbLite instance}

\item{by}{how to split the genes}
}
\value{
a GRangesList

a GRangesList
}
\description{
A parent class for lightweight Ensembl and RepBase/RepeatMasker annotations
}
\section{Methods (by generic)}{
\itemize{
\item \code{transcriptsBy}: get transcripts by gene, promoter, tx_biotype, gene_biotype, or biotype_class

\item \code{genesBy}: get genes by gene_biotype or biotype_class
}}

