% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxDbLite-class.R, R/TxDbLite-methods.R
\docType{class}
\name{TxDbLite-class}
\alias{TxDbLite-class}
\alias{EnsDbLite-class}
\alias{RepDbLite-class}
\alias{ErccDbLite-class}
\alias{ArrayControlDbLite-class}
\alias{dbconn,TxDbLite-method}
\alias{show,TxDbLite-method}
\alias{metadata,TxDbLite-method}
\alias{transcriptsBy,TxDbLite-method}
\alias{genesBy}
\alias{genesBy,TxDbLite-method}
\alias{genesBy}
\alias{TxDbLite-method}
\title{A parent class for lightweight Ensembl and RepBase/RepeatMasker annotations}
\usage{
\S4method{dbconn}{TxDbLite}(x)

\S4method{show}{TxDbLite}(object)

\S4method{metadata}{TxDbLite}(x, ...)

\S4method{transcriptsBy}{TxDbLite}(x, by = c("gene", "promoter", "tx_biotype",
  "gene_biotype", "biotype_class"))

genesBy(x, by = c("gene_biotype", "biotype_class"), ...)

\S4method{genesBy}{TxDbLite}(x, by = c("gene_biotype", "biotype_class"))
}
\arguments{
\item{x}{TxDblite SQL-lite database instance}

\item{object}{TxDbLite SQL-lite annotation database}

\item{...}{additional parameters releated to annotation database}

\item{by}{how to split the genes}
}
\value{
a GRangesList

a GRangesList
}
\description{
A parent class for lightweight Ensembl and RepBase/RepeatMasker annotations

A parent class for Ensembl based annotations

Repeat Parent Annotation Class

Ercc Based Annotation Class

ArrayControl Based Annotation Class

get transcripts by gene, promoter, tx_biotype, gene_biotype, or biotype_class

Generic for querying genes

get genes by gene_biotype or biotype_class
}
\section{Slots}{

\describe{
\item{\code{con}}{a DBI Connection}

\item{\code{tables}}{a list of annotation entries}
}}

