% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxDbLite.R
\name{TxDbLite}
\alias{TxDbLite}
\title{Create a TxDbLite object (usually called by subclass constructors)}
\usage{
TxDbLite(x, path = ".", ...)
}
\arguments{
\item{x}{an annotation instance of TxDbLite}

\item{path}{where it lives (default: ".")}

\item{...}{additional parameters}
}
\value{
a TxDbLite object
}
\description{
Create a TxDbLite object (usually called by subclass constructors)
}
