% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TxDbLite.R
\name{TxDbLite}
\alias{TxDbLite}
\title{Create a TxDbLite object (usually called by subclass constructors)}
\usage{
TxDbLite(x, path = ".", ...)
}
\arguments{
\item{x}{name of the sqlite file whence the TxDbLite should be created}

\item{path}{where it lives (default: ".")}
}
\value{
a TxDbLite object
}
\description{
Create a TxDbLite object (usually called by subclass constructors)
}

