% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{utils}
\alias{colAsVector}
\alias{getFastaStub}
\alias{getOrgDetails}
\alias{getOrganismAbbreviation}
\alias{getSupportedAbbreviations}
\alias{getTxDbLiteName}
\alias{utils}
\title{utility functions}
\format{An object of class \code{NULL} of length 0.}
\usage{
utils

getFastaStub(fastaFile)

colAsVector(x, y)

getSupportedAbbreviations()

getOrganismAbbreviation(organism)

getOrgDetails(organism)

getTxDbLiteName(fastaFile)
}
\description{
utility functions
}
\section{Functions}{
\itemize{
\item \code{getFastaStub}: get the "stub" of a FASTA filename (no .fa, no .fasta, no .gz)

\item \code{colAsVector}: get a column of a data.frame _complete_with_rownames_ as a vector

\item \code{getSupportedAbbreviations}: supported organism abbreviations

\item \code{getOrganismAbbreviation}: Get the abbreviation for an organism (among those with which we're familiar)

\item \code{getOrgDetails}: helper fn that handles a number of annoying tasks using saved data

\item \code{getTxDbLiteName}: get the name of the package/sqlite file for a FASTA-based annotation
NOTE: as of 1.9.25, this is based on DbType.Org.Version
}}
\keyword{datasets}

