% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{utils}
\alias{colAsVector}
\alias{getFastaStub}
\alias{getOrgDetails}
\alias{getOrganismAbbreviation}
\alias{getSupportedAbbreviations}
\alias{getTxDbLiteName}
\alias{idStub}
\alias{utils}
\title{utility functions}
\format{An object of class \code{NULL} of length 0.}
\usage{
utils

idStub(x)

getFastaStub(fastaFile)

colAsVector(x, y)

getSupportedAbbreviations(how = c("abbreviation", "reactome"))

getOrganismAbbreviation(organism, how = c("abbreviation", "reactome"))

getOrgDetails(organism)

getTxDbLiteName(fastaFile)
}
\arguments{
\item{x}{a string to try and pop the ENSEMBL identifier off of}
}
\value{
if everything works out properly, the ENS* identifier (without .N)
}
\description{
utility functions
}
\section{Functions}{
\itemize{
\item \code{idStub}: needed from Ens83 onwards :-/ and maybe will need to generalize as well

\item \code{getFastaStub}: get the "stub" of a FASTA filename (no .fa, no .fasta, no .gz)

\item \code{colAsVector}: get a column of a data.frame _complete_with_rownames_ as a vector

\item \code{getSupportedAbbreviations}: supported organism abbreviations

\item \code{getOrganismAbbreviation}: Get the abbreviation for an organism (among those with which we're familiar)

\item \code{getOrgDetails}: helper fn that handles a number of annoying tasks using saved data

\item \code{getTxDbLiteName}: get the name of the package/sqlite file for a FASTA-based annotation
NOTE: as of 1.9.25, this is based on DbType.Org.Version
}}
\keyword{datasets}

