% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{utils}
\alias{utils}
\alias{idStub}
\alias{getFastaStub}
\alias{colAsVector}
\alias{getSupportedAbbreviations}
\alias{getOrganismAbbreviation}
\alias{getOrgDetails}
\alias{getTxDbLiteName}
\title{utility functions}
\format{An object of class \code{NULL} of length 0.}
\usage{
utils

idStub(x)

getFastaStub(fastaFile)

colAsVector(x, y)

getSupportedAbbreviations(how = c("abbreviation", "reactome"))

getOrganismAbbreviation(organism, how = c("abbreviation", "reactome"))

getOrgDetails(organism)

getTxDbLiteName(fastaFile)
}
\arguments{
\item{x}{a string to try and pop the ENSEMBL identifier off of}

\item{fastaFile}{a FASTA file of interest 
get the "stub" of a FASTA filename (no .fa, no .fasta, no .gz)}

\item{y}{a column of a data frame
get a column of a data.frame _complete_with_rownames_ as a vector}

\item{how}{method for finding supported species abbreviations 
supported organism abbreviations}

\item{organism}{a familiar organism 
Get the abbreviation for an organism (among those with which we're familiar)}
}
\value{
if everything works out properly, the ENS* identifier (without .N)
}
\description{
utility functions

Finds the prefix name

helper fn that handles a number of annoying tasks using saved data

get the name of the package/sqlite file for a FASTA-based annotation
NOTE: as of 1.9.25, this is based on DbType.Org.Version
}
\keyword{datasets}
