% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{utils}
\alias{createAnnotationPackage}
\alias{getAnnotationType}
\alias{getFastaStub}
\alias{getOrganismAbbreviation}
\alias{getTxDbLiteName}
\alias{utils}
\title{utility functions}
\format{\preformatted{ NULL
}}
\usage{
utils

getFastaStub(fastaFile)

getOrganismAbbreviation(organism)

getTxDbLiteName(fastaFile)

getAnnotationType(fastaFile)

createAnnotationPackage(fastaFile, author)
}
\arguments{
\item{fastaFile}{the filename}

\item{author}{the author (with email address)}
}
\value{
the name of the annotation package, or NULL if uncertain how to do it
}
\description{
utility functions
}
\section{Functions}{
\itemize{
\item \code{getFastaStub}: get the "stub" of a FASTA filename (no .fa, no .fasta, no .gz)

\item \code{getOrganismAbbreviation}: Get the abbreviation for an organism (among those with which we're familiar)

\item \code{getTxDbLiteName}: get the name of the package/sqlite file for a FASTA-based annotation

\item \code{getAnnotationType}: figure out what type of annotation package to create

\item \code{createAnnotationPackage}: create an annotation package for a FASTA file (try to figure out what kind)
}}
\keyword{datasets}

